Feature best delivery driver (#1)
* test protected

* update

* accept to deliver order

* assign drivers to deliver a order

* fixed some stupid errors

+++ restaurants/src/main/java/com/zeco/restaurants/model/Orders.java
@@ -51,10 +51,10 @@ public class Orders {
     private String deliveryInstructions;
 
     @Column(name = "delivery_latitude")
-    private Float deliveryLatitude;
+    private Double deliveryLatitude;
 
     @Column(name = "delivery_longitude")
-    private Float deliveryLongitude;
+    private Double deliveryLongitude;
 
     @OneToMany(mappedBy = "orderID", orphanRemoval = true, cascade = CascadeType.ALL)
     private List<OrderDishes> orderDishesList = new ArrayList<>();
